+++ src/time_steppers/cc_factory_time_stepper.h
@@ -1,0 +1,8 @@
+// Include header and tpl.cpp files implementing templates. We may use
+// this file to force the instantiation of specific templates
+#ifndef CCFACTORYTIMESTEPPERFORODES_H
+#define CCFACTORYTIMESTEPPERFORODES_H
+#include "cc_factory_time_stepper.tpl.h"
+#include "cc_factory_time_stepper.tpl.cpp"
+#endif // #ifndef CCFACTORYTIMESTEPPERFORODES_H
+
