+++ src/time_steppers/cc_factory_time_stepper.tpl.h
@@ -1,8 +1,8 @@
-#ifndef CCFACTORYTIMESTEPPERFORODES_H
-#define CCFACTORYTIMESTEPPERFORODES_H
+#ifndef CCFACTORYTIMESTEPPER_TPL_H
+#define CCFACTORYTIMESTEPPER_TPL_H
 
 // Include the integration methods for odes (time steppers)
-#include "ac_time_stepper_for_odes.h"
+#include "ac_time_stepper.h"
 #include "cc_euler_method.h"
 #include "cc_runge_kutta_4_method.h"
 #include "cc_backward_euler_predictor_corrector_method.h"
@@ -16,45 +16,46 @@
 namespace scicellxx
 {
 
- /// @class CCFactoryTimeStepperForODEs cc_factory_time_stepper_for_odes.h
+ /// @class CCFactoryTimeStepper cc_factory_time_stepper.h
 
  /// This class implements a factory for the integration methods for
  /// ODEs (time steppers)
- class CCFactoryTimeStepperForODEs
+ template<class EQUATIONS_TYPE>
+ class CCFactoryTimeStepper
  {
   
  public:
   
   /// Empty constructor
-  CCFactoryTimeStepperForODEs();
+  CCFactoryTimeStepper();
   
   /// Empty destructor
-  virtual ~CCFactoryTimeStepperForODEs();
+  virtual ~CCFactoryTimeStepper();
   
   /// Returns the specified time stepper (integration method)
-  ACTimeStepperForODEs* create_time_stepper(std::string time_stepper_name);
+  ACTimeStepper<EQUATIONS_TYPE>* create_time_stepper(std::string time_stepper_name);
   
  protected:
   
   /// Copy constructor (we do not want this class to be
   /// copiable). Check
   /// http://www.learncpp.com/cpp-tutorial/912-shallow-vs-deep-copying/
-  CCFactoryTimeStepperForODEs(const CCFactoryTimeStepperForODEs &copy)
+  CCFactoryTimeStepper(const CCFactoryTimeStepper &copy)
    {
-    BrokenCopy::broken_copy("CCFactoryTimeStepperForODEs");
+    BrokenCopy::broken_copy("CCFactoryTimeStepper");
    }
   
   /// Assignment operator (we do not want this class to be
   /// copiable. Check
   /// http://www.learncpp.com/cpp-tutorial/912-shallow-vs-deep-copying/
-  void operator=(const CCFactoryTimeStepperForODEs &copy)
+  void operator=(const CCFactoryTimeStepper &copy)
    {
-    BrokenCopy::broken_assign("CCFactoryTimeStepperForODEs");
+    BrokenCopy::broken_assign("CCFactoryTimeStepper");
    }
  
  };
 
 }
  
-#endif // #ifndef CCFACTORYTIMESTEPPERFORODES_H
+#endif // #ifndef CCFACTORYTIMESTEPPER_TPL_H
